client/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-neopixel==6.3.12
adafruit-circuitpython-neopixel-spi==1.0.10
adafruit-circuitpython-pixelbuf==2.0.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2024.8.30
charset-normalizer==3.4.0
dbus-python==1.2.16
idna==3.10
numpy==1.19.5
opencv-python==4.10.0.84
psutil==5.9.5
pycairo==1.16.2
pycurl==7.43.0.6
pyftdi==0.55.4
PyGObject==3.42.2
pyserial==3.5
python-apt==2.2.1
python-periphery==2.4.1
pyusb==1.2.1
requests==2.32.3
rknn-toolkit-lite2 @ file:///home/rock/rknn-toolkit2/rknn_toolkit_lite2/packages/rknn_toolkit_lite2-1.5.0-cp39-cp39-linux_aarch64.whl
rpi-ws281x==5.0.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sysv-ipc==1.1.0
typing-extensions==4.12.2
urllib3==1.26.5
